fix: friends 패키지 util 클래스 리팩토링
