+++ src/main/java/gravit/code/friend/infrastructure/strategy/NicknameSearchStrategy.java
@@ -2,7 +2,8 @@ package gravit.code.friend.infrastructure.strategy;
 
 import gravit.code.friend.dto.SearchPlan;
 import gravit.code.friend.infrastructure.sql.select.FriendsNicknameSearchQuerySql;
-import gravit.code.friend.util.QueryNormalizeUtil;
+import gravit.code.friend.support.HandleNormalize;
+import gravit.code.friend.support.NicknameNormalize;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -24,7 +25,7 @@ public class NicknameSearchStrategy implements FriendsSearchStrategy {
             int page,
             int size
     ) {
-        String cleanText = QueryNormalizeUtil.nicknameNormalize(queryText);
+        String cleanText = NicknameNormalize.nicknameNormalize(queryText);
 
         if(cleanText.isEmpty()){
             return SearchPlan.empty();
